refactor: rafacored hooks, const for additional scroll value
