--- packages/tabs/src/components/tabs/mobile/index.module.css
@@ -1,3 +1,0 @@
-.wrapper {
-    margin: 0 var(--gap-m-neg);
-}
