+++ packages/tabs/src/useFullWidthScroll.ts
@@ -9,47 +9,24 @@ export const useFullWidthScroll = (isFullScroll: boolean) => {
 
     const handleScroll = (e: UIEvent<HTMLDivElement>) => {
         const node = e.currentTarget;
-        const containerNode = containerRef.current;
 
-        if (containerNode) {
-            if (node.scrollLeft) {
-                setIsLeftOut(true);
-            } else {
-                setIsLeftOut(false);
-            }
-            if (node.scrollLeft + node.clientWidth >= node.scrollWidth - 2) {
-                setIsRightOut(false);
-            } else {
-                setIsRightOut(true);
-            }
-        }
+        setIsLeftOut(node.scrollLeft !== 0);
+        setIsRightOut(node.scrollLeft + node.clientWidth >= node.scrollWidth);
     };
 
     useEffect(() => {
-        const scrollableContainerNode = scrollableContainerRef.current;
-        const containerNode = containerRef.current;
+        const scrollableNode = scrollableContainerRef.current || containerRef.current;
 
-        if (containerNode && isFullScroll) {
-            if (scrollableContainerNode) {
-                const { scrollWidth, clientWidth } = scrollableContainerNode;
-                if (scrollWidth > clientWidth) {
-                    setIsRightOut(true);
-                }
-            } else {
-                const { scrollWidth, clientWidth } = containerNode;
-                if (scrollWidth > clientWidth) {
-                    setIsRightOut(true);
-                }
-            }
+        if (scrollableNode && isFullScroll) {
+            const { scrollWidth, clientWidth } = scrollableNode;
+            if (scrollWidth > clientWidth) setIsRightOut(true);
         }
     }, [isFullScroll]);
-
-    const isHandleScroll = isFullScroll && scrollableContainerRef.current;
 
     return {
         containerRef,
         scrollableContainerRef,
-        handleScroll: isHandleScroll ? handleScroll : undefined,
+        handleScroll,
         isRightOut,
         isLeftOut,
     };
